+++ ui/src/main/java/io/snabble/sdk/ui/payment/creditcard/datatrans/ui/DatatransViewModel.kt
@@ -86,7 +86,6 @@ internal class DatatransViewModel(
         }
         options.appCallbackScheme = "snabble"
         options.isTesting = isTesting
-        options.useCertificatePinning = true
     }
 
     fun errorHandled() {